Jupyter REPL: automatic show last variable's value
Changed the functions `acg/eval-with` and the ones that were being used
with it to automatically read the last line of the code that is being
sent to the REPL and, if there is a variable assignment, append a
newline to the code containing the variable so that the REPL will show
the variable's value.

Extra: added handling of Pandas Series in the "open in external app"
function for the Jupyter mode.