Custom Window Split behavior
Customized the default split window behavior of Emacs. Originally, Emacs
would always split the current window if it's the only one on its frame.
We removed this so that now the thresholds are always respected.

With this fix we were able to restablish the default display buffer
functions for help, magit, and similar buffers.