Better custom Split Window behavior
The previous customization of split window was making default functions
(of Emacs and other packages) create new frames in an undesirable
manner (when insert a link in Org mode, for instance).

In this commit we improve the customization by allowing Emacs to split
the current window even when thresholds are not met, but then set the
default behavior to not split in that case. This prevents the creation
of new frames still allowing the desired default behavior to happen for
the majority of cases.