Base custom Font config on $DISPLAY env var
Refactored the custom font configuration from the wsl specific configs
to the UI file. Applied a conditional triggering of the custom font
based on the environment variable "DISPLAY", that is usually set when
using X11 Forwarding to display graphic applications remotely, which is
exactly when we want to apply custom font configurations.