Dissociate tab-line from buffer history
Created a custom `tab-line-tabs-function` to have the tab list order do
not be tied to the history of buffers opened in this window. The idea is
that we want to preserve the position of the tabs in the order in which
they were opened (or in the order we want).