Undo-tree: update custom funcs and make ESC abort
Updated the custom functions I had created previously with their actual
behavior description. My previous modifications had a side effect I did
not notice: they were not only keeping the previous region state but
also not moving the cursor when undoing, which is undesired.

Also, modified the binding of the ESC key to quit and abort the
undo-tree-visualizer (making the buffer go back to its original state --
when the user started the visualizer in the first place).