Standardize edit of src code blocks
Created a standard way that Emacs will handle editing, committing and
abortion of source code blocks for both Org and Markdown modes.

The idea is to have the same keyboard shortcuts behaving the same way
for the two modes. Also, to have a confirmation question when one tries
to kill an edit src buffer after it has been modified.