Fix fixed-pitch font not fixed in variable-pitch
When Emacs was evaluating the definition of the fixed-pitch font and a
frame was not loaded (such as when initializing Emacs), the default font
from which the fixed-pitch was inheriting its family attribute was not
yet set to a monospace font. This commit fixes it by making the setting
of the variable happen when the mode is enabled.