Fix move-line clumping last two lines
When operating on the last lines of a buffer, the command `move-line`
would clump the last two lines together in case the last line of a
buffer didn't end with a newline (final newline).

Here we fix it by ensuring we add the final newline if necessary when
the command is run for the first time.