Vertico: fix identification of browsing dir
The appropriate context directory was not being correctly identified by
the function `acg/vertico--candidate-dir`, due to the return value of
`vertico--candidate` being the empty string when no candidate was
selected.

Since still in this case the user's intention in changing to another
command would be to do so in the latest browsing directory, this case
has been included in the function.