Restore default M-x command
Here we restore the M-x keybinding to be `execute-extended-command`. The idea
is now to get rid of "cut" (clipboard-kill-region), as this can easily be
performed by copying and deleting.

In order to allow for copying and deleting, we disable Emacs's default copy
behavior of deselecting the region, and we preserve the region as is.